Binary classifier take observation weights; classifiers vote to initialize dIdx
Can optionally set the modelOpts.balance field with a string indicating the field in labels.windows over which you want to balance the classifier observation weights.
If modelOpts.balance is set, observation weights are also balanced by class and mouse(subject).
If you indicate the number of supervised factors to initialize, but not the indicies, those indicies are now chosen by having all of the classifiers vote. The indicies with the most votes win. Ties are broken based on the order that the classifiers are given.