First commit for stochastic-by-frequency optimization.
At each iteraction a batch of frequencies is randomly selected, and gradients are only applied for those frequencies.
This applies to both score learning and kernel learning.
By default, stochastic-by-frequency optimization is on with batch size of 8, controlled by the fStochastic and fBatchSize options, respectively.
Stochastic-by-frequency optimization is turned off in projectCSFA.